Rename `BuildFilePlatformFilter` to `PlatformFilter`
