+++ Sources/TuistCoreTesting/Graph/GraphDependencyReference+TestData.swift
@@ -70,7 +70,7 @@ extension GraphDependencyReference {
     public static func testProduct(
         target: String = "Target",
         productName: String = "Target.framework",
-        platformFilter: BuildFilePlatformFilter = .ios
+        platformFilter: PlatformFilter = .ios
     ) -> GraphDependencyReference {
         GraphDependencyReference.product(
             target: target,
